+++ container/common/templates/role/defaults/main.j2.yml
